pyproject.toml
[project]
name = "apocalypto_env"
version = "1.0.0"
description = "Meta x HuggingFace OpenEnv Hackathon India submission"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi",
    "uvicorn",
    "pydantic>=2.0.0",
    "openai"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
